refactor(9b-3): decompose FooterSettingsCard — extract SocialLinkEditForm + SocialLinkRow
Card dropped 502 -> 294 lines following the proven Bundle 8 template
(presentational sub-components receive props + callbacks, own no data fetching).

NEW sub-components (both in components/admin/settings/):
- SocialLinkEditForm.tsx (173 lines) — the icon-picker + label/url inputs + save/cancel
  actions. This was ALREADY a separate function component in the original file; moved to
  its own module along with its ICON_OPTIONS/IconKey/EditState/emptyEdit dependencies
  (which only it uses). Exports the types so the parent can still construct EditState.
- SocialLinkRow.tsx (89 lines) — the display-mode row (icon + label/url + edit/delete
  action buttons), previously inlined in the links.map() body.

Stays in FooterSettingsCard: the copyright-text editor (tightly coupled to its own state +
save handler), the social-links list orchestration (load/startEdit/startNew/save/remove
mutations + the empty state + the Add button + the inline edit/new form wiring), the
collapsible header, and the accordion shell.

Added 10 net new smoke tests (SocialLinkEditForm 5, SocialLinkRow 3, emptyEdit helper 2).
The 19 existing FooterSettingsCard tests pass unchanged — including the edit/delete/icon-
change tests that exercise both extracted components. Frontend 1628 pass / 0 fail (1618 +
10). tsc + prettier + lint clean.
